♻️ : 동전 목록 재사용 가능하도록 컴포넌트 분리
