+++ src/Components/@Shared/CoinList/styles.scss
