+++ src/Components/@Shared/CoinList/template/index.html
@@ -1,0 +1,12 @@
+<section id="holding-amount-table-section" class="table-section">
+  <table id="holding-amount-table" class="table">
+    <caption>자판기가 보유한 동전</caption>
+    <thead>
+      <tr>
+        <th>동전</th>
+        <th>개수</th>
+      </tr>
+    </thead>
+    <tbody></tbody>
+  </table>
+</section>;
