--- src/Components/CoinList/template/index.html
@@ -1,29 +1,0 @@
-<section id="holding-amount-table-section" class="table-section">
-  <table id="holding-amount-table" class="table">
-    <caption>자판기가 보유한 동전</caption>
-    <thead>
-      <tr>
-        <th>동전</th>
-        <th>개수</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr>
-        <td>500원</td>
-        <td>0개</td>
-      </tr>
-      <tr>
-        <td>100원</td>
-        <td>0개</td>
-      </tr>
-      <tr>
-        <td>50원</td>
-        <td>0개</td>
-      </tr>
-      <tr>
-        <td>10원</td>
-        <td>0개</td>
-      </tr>
-    </tbody>
-  </table>
-</section>
